+++ app/AIStudioLocal/Domain/ExportTemplate.swift
@@ -13,9 +13,3 @@ public struct ExportTemplate: Codable, Identifiable, Equatable {
         self.element = ContinuityElement(type: .exportTemplate, name: name, promptBlock: promptBlock)
     }
 }
-
-extension ExportTemplate {
-    public static var mock: ExportTemplate {
-        ExportTemplate(name: "TikTok Vertical", promptBlock: "9:16 vertical export, 1080x1920, H.264")
-    }
-}